Move ObjectEntry.h to correct include directory
