+++ blort_ros/include/blort_ros/ObjectEntry.h
