+++ blort_ros/src/gltracker.cpp
@@ -364,7 +364,7 @@ void GLTracker::update()
             // although the implementation would allow it, at several places, lacks this at several other.
             tracker.getModelMovementState(model_ids[i], movements[i]);
             tracker.getModelQualityState(model_ids[i], qualities[i]);
-            ROS_INFO_STREAM("GLTracker::update: the tracked model for " << model_names_[i] << " has set quality to " << qualities[i]);
+            ROS_INFO_STREAM("GLTracker::update: the tracked model for " << objects_[i].name << " has set quality to " << qualities[i]);
             tracker.getModelConfidenceState(model_ids[i], tracking_confidences[i]);
 
             switch(tracking_confidences[i])
